backend/requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
requests
gTTS
Pillow
numpy>=1.26.0
scikit-learn>=1.7.0
pandas>=2.0.0
scipy>=1.11.0